requirements.txt
streamlit==1.31.0
opencv-python-headless==4.8.0
pytesseract==0.3.10
pandas==2.0.3
openpyxl==3.1.2
Pillow==10.0.0 